requirements.txt
fastapi
uvicorn
celery
redis
openai
sqlalchemy
psycopg2-binary
python-dotenv
requests
slowapi